Basic Split View
Implemented the split table view. Basically once the data is parsed and
the arrays are evened out this will display the data as a bunch of
UITableViewCells that match up just like they would on the GitHub split
view.

Still need to finish up the line numbers and clean up the code but the
basic idea of how this will work is there.